Moved syscalls from syscall.h to respective files and did test for sbush

+++ include/unistd.h
@@ -3,7 +3,7 @@
 
 #include <sys/defs.h>
 
-int close(int fd);
+int8_t close(int fd);
 ssize_t read(int fd, void *buf, size_t count);
 ssize_t write(int fd, const void *buf, size_t count);
 int unlink(const char *pathname);
